As per tweets: https://twitter.com/indutny/status/557499042243289088

It seems new handling for upgrade was added at some point, which makes the tests break. Unfortunately I don't have time to work out how to make the tests pass again.

SOURCE should essentially be handled like Upgrade, I believe, and switched to data/tcp mode after parsing headers.

/cc @indutny @bnoordhuis