@indutny 2015-01-20
I'm quite sure that it doesn't affect it, but nevertheless.

May I ask you to run the benchmark before and after the change and post the results here?