@miksago 2015-01-20
This changeset would also affect the decision of the Icecast project to move to using http-parser: https://trac.xiph.org/ticket/2057