@miksago 2015-01-20
With patch:

```
Benchmark result:
Took 12.565303 seconds to run
397921.156250 req/sec
```

Without:

```
Benchmark result:
Took 14.347905 seconds to run
348482.937500 req/sec
```

I'm not particularly sure why it took longer without the patch. Ran a second time, and got:

```
Benchmark result:
Took 11.664329 seconds to run
428657.343750 req/sec
```