@indutny 2015-01-20
Ok, I see. Are tests passing on your machine? It seems that CI is not happy with it.