@miksago 2015-01-20
My machine is under pretty heavy load (memory, disk) and may have a hard drive failure happening soon.