@miksago 2015-01-20
No, see my note in the comment about the fix_upgrade_message failing. I don't know how to fix straight away, and can't put any more time in right now