@jasnell 2015-10-26
I will note that `SOURCE` is not in the official list of registered methods: http://www.iana.org/assignments/http-methods/http-methods.xhtml