@miksago 2015-10-27
Indeed, it is not, although, it also turns out the icecast protocol isn't exactly HTTP anyway. 

The reason for this change was a suggestion to the icecast dev-team to try using this http 
parser instead of rewriting their own. 

On 26 Oct 2015, at 9:57 pm, James M Snell <notifications@github.com> wrote:

> I will note that SOURCE is not in the official list of registered methods: http://www.iana.org/assignments/http-methods/http-methods.xhtml
> 
> �
> Reply to this email directly or view it on GitHub.
>