@jasnell 2015-10-27
While that kind of reuse is definitely a noble goal, I'm not sure it's the right thing to do in http-parser right now. The [HTTP WG](http://github.com/nodejs/http) will be taking another look at HTTP support in core. Part of that will be revisiting the http-parser and exploring what additional changes should be made -- including support for arbitrary HTTP methods as opposed to a limited fixed list. Up to this point, the philosophy has been to support only methods that are included in the official registry.