@ThisIsMissEm 2020-09-21
Looks like this may have been fixed by #431 and #407, though those didn't treat the request as being a switching protocols type request, so, no upgrade event.